Fix host CMake build: include paths + missing sources
The documented `cmake && make && ctest` flow never compiled:
  - target_include_directories only listed `include src`, so the first
    engine TU failed on `#include "mnv_blake2s.h"` (lives in src/security).
  - MNV_SOURCES omitted mnv_lut.c and mnv_outauth.c, which the engine
    and the host test both reference -> link failure.

Add every src/ subdir to the include path and the two missing TUs to the
library. This is also the first time the engine/arch sources compile under
CMake; fix the one -Wall -Wextra warning that surfaced (signed/unsigned
compare in mnv_blake2s_verify).

Red-team: `cmake -S . -B build -DMNV_TARGET=host && cmake --build build`
is now warning-free and `ctest` passes 32/32.

+++ src/security/mnv_blake2s.c
@@ -163,6 +163,6 @@ mnv_status_t mnv_blake2s_verify(const uint8_t *key,    uint8_t  keylen,
     mnv_blake2s_mac(key, keylen, data, datalen, computed);
     uint8_t diff = mnv_ct_compare(computed, expected_mac, MNV_BLAKE2S_DIGEST_SIZE);
     volatile uint8_t *vp = (volatile uint8_t *)computed;
-    for (int i = 0; i < MNV_BLAKE2S_DIGEST_SIZE; i++) vp[i] = 0;
+    for (size_t i = 0; i < MNV_BLAKE2S_DIGEST_SIZE; i++) vp[i] = 0;
     return (diff == 0) ? MNV_OK : MNV_ERR_TAMPER;
 }
